增加令牌刷新功能，取消验证token时判断ip是否一致（原因：会导致如果使用网络代理会验证token失败从而导致登录失效）

+++ src/main/java/com/ibaiq/common/constants/Constants.java
@@ -8,7 +8,7 @@ package com.ibaiq.common.constants;
 public interface Constants {
 
     // 登录过期时间
-    long LOGIN_EXPIRED = 60 * 60 * 24;
+    int LOGIN_EXPIRED = 60 * 60 * 24;
 
     String REDIS_PREFIX_LOGIN_EXPIRED = Constants.REDIS_PREFIX_CONFIG + "sys.login.expired";
